@@ -1,7 +1,7 @@
 package com.kienht.bubble_picker
 
 import android.content.Intent
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import kotlinx.android.synthetic.main.activity_main.*
 
@@ -16,15 +16,15 @@ class MainActivity : AppCompatActivity() {
 
         setContentView(R.layout.activity_main)
 
-        button_sync_activity.setOnClickListener({
+        button_sync_activity.setOnClickListener {
             val intent = Intent(this@MainActivity, SyncActivity::class.java)
             startActivity(intent)
-        })
+        }
 
-        button_async_activity.setOnClickListener({
+        button_async_activity.setOnClickListener {
             val intent = Intent(this@MainActivity, AsyncActivity::class.java)
             startActivity(intent)
-        })
+        }
     }
 
 }
